+++ packages/yak_runner/lib/src/delegates/test/delegate/nullary.dart
@@ -1,8 +1,0 @@
-import 'dart:async';
-
-/// a delegate for compatibility with [https://pub.dev/packages/yak_test]
-/// or 3rd party custom implementation
-mixin RunnerTestDelegate<T> {
-  /// `call` will run the test
-  void call(FutureOr<T> result);
-}