+++ packages/yak_runner/lib/src/mixin/test/nullary.dart
@@ -1,9 +1,0 @@
-import '../../all.dart';
-
-/// a `mixin` on `RunnerBase`
-mixin RunnerTestMixin<T> on RunnerBase<T> {
-  /// is a `RunnerTestDelegate Function(RunnerTestFactoryDelegate)`
-  /// type `T` is passed by the `Object` to the test to `RunnerTestDelegate`
-  RunnerTestDelegate<T> buildTest(RunnerTestFactoryDelegate testFactory) =>
-      testFactory<T>();
-}