+++ packages/yak_runner/test/mixin/test/test_factory/nullary.dart
@@ -1,9 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-
-import '../test_delegate/nullary.dart';
-
-class FakeNullaryTestFactory implements RunnerTestFactoryDelegate {
-  const FakeNullaryTestFactory();
-  @override
-  RunnerTestDelegate<T> call<T>() => const FakeNullaryTest();
-}