+++ packages/yak_runner/test/mixin/test/test_factory/unary.dart
@@ -1,9 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-
-import '../test_delegate/unary.dart';
-
-class FakeUnaryTestFactory implements UnaryRunnerTestFactoryDelegate {
-  const FakeUnaryTestFactory();
-  @override
-  UnaryRunnerTestDelegate<T, S> call<T, S>() => const FakeUnaryTest();
-}