+++ packages/yak_runner/test/mocks/all.dart
@@ -1,2 +1,1 @@
-export 'delegates/all.dart';
 export 'json/all.dart';