+++ packages/yak_test/test/sync/nullary_test.dart
@@ -1,9 +1,0 @@
-import 'package:yak_runner/yak_runner.dart';
-import 'package:yak_test/yak_test.dart';
-
-final runner = Runner(() => 0);
-final _test = runner.buildTest(
-  runner.buildTestFactory('unary test'),
-);
-
-void main() async => _test(0);